cs329s_waymo_object_detection/streamlit_app/requirements.txt
streamlit
Pillow
numpy
pandas
opencv-python
matplotlib
glob2
cs329s-waymo-object-detection==0.1.0